+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes for Craft Generator
 
-## Unreleased
+## 1.1.0 - 2023-01-06
 - Added the Twig Extension generator (`twig-extension'). ([#8](https://github.com/craftcms/generator/discussions/8))
 - Added support for generating components within arbitrary source paths, via a new `--path` option. 
 - Improved the web controller scaffolding.